fix(pricing): accept rabatt_prozent als float fuer praezise End-Rabatte
Shop-API liefert effective_rebate mit Dezimalstellen (z.B. 60.9). Der
(int)-Cast verlor diese Praezision -> rechnerischer Effektivpreis wich
um bis zu 0.03 EUR vom price_net ab und triggerte die Diff-Warnung.

Mit (float) erkennt der Calculator 60.9 als 60.9 und matched price_net
exakt. Validiert an 12 Live-Tarifen aus mb-sandbox (0/12 Warnungen).

+++ src/Pricing/Calculator.php
@@ -105,7 +105,7 @@ final class Calculator
         $features        = $t['detail_features'] ?? $t['features'] ?? [];
         $groupName       = (string) ($t['group_name'] ?? '');
 
-        $rabattPct       = (int) ($t['rabatt_prozent'] ?? 0);
+        $rabattPct       = (float) ($t['rabatt_prozent'] ?? 0);
         $startguthaben   = (float) ($t['startguthaben'] ?? -1);
         $gratisMonate    = (int) ($t['gratis_monate'] ?? -1);
         $laufzeitMonate  = (int) ($t['laufzeit_monate'] ?? 0);
